+++ src/index.ts
@@ -5,7 +5,7 @@
  * @packageDocumentation
  */
 
-export const VERSION = '2.1.0';
+export const VERSION = '2.2.0';
 
 // Infrastructure exports
 export { HttpMethod } from './infrastructure';