+++ src/theme/index.ts
@@ -5,3 +5,4 @@ export * from './ThemeManager';
 export * from './ThemeProvider';
 export * from './useTheme';
 export * from './useThemedStyles';
+export * from './useTextStyle';